cleaned up Python code and prepped for next feature
