+++ control/utils.py
@@ -1,0 +1,25 @@
+"""
+This file contains utility/helper functions used across multiple modules/files
+"""
+
+import socket
+
+"""
+	Create a server socket and listen for an incoming connection
+	Returns the socket associated with the connection
+"""
+def openConnection(portNum):
+	print(f"Listening for connection on port {portNum} ...")
+
+	# create an IPv4 TCP server socket and set it up to listen
+	serversocket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+	serversocket.bind(('', portNum))
+	serversocket.listen()
+
+	# listen for incoming connections and accept them
+	(clientsocket, address) = serversocket.accept()
+	print("Connection established")
+
+	# close the server socket and return the socket associated with the connection
+	serversocket.close()
+	return clientsocket
